🩹 Fix: type of `*::oldest_record_ts` fields from `f64` to `i64` [#15]

+++ src/model/labs/leagueflow.rs
@@ -33,7 +33,7 @@ impl AsRef<LabsLeagueflowResponse> for LabsLeagueflowResponse {
 pub struct LabsLeagueflow {
     /// The timestamp of the oldest record found.
     #[serde(rename = "startTime")]
-    pub oldest_record_ts: f64,
+    pub oldest_record_ts: i64,
     /// The points in the chart.
     ///
     /// - 0: The timestamp offset.
